Working Shortest path from root node
Created a general tree using chat gpt, modified the code for the
requirements of the program. The code is inside the
GeneralTreeCreationReference.

Did not properly comment out the code, some parts of the code is
confusing and questionable. Will have to check and understand it first.
Before submitting this as a done project.