Working : Finding Shortest Distance Between nodes
Added a new method to find the shortest distance between two nodes.
Generated by ChatGPT, methods used/referenced by ChatGPT have some
alterations inside the class.

I renamed some variables to be more descriptive, longer and more verbose
than what is needed but I really need it to understand exactly what is
going on.

The tree generated for testing purposes is not from user input, the
nodes to be searched for within the tree are the only ones which are
from user.