WORKING: Use user input for Creating general tree
Program can now accept a string from user which represents the numbers
for the nodes inside the tree. The first of the pair of numbers, or the
first number in the left side is the root node of the general tree.
Then the number that follows is the child of the root node.

The pattern goes like parent - child, the numbers should be separated by
space. The space is used as the regex for the program.