WORKING:Multiple query for Node distance
Code needs to be commented, which I don't know where to put
said comments.

Added a function to create an set of unordered pair of int numbers
from a String provided by user.

Removed the test fucntions, one for finding the shortest distance
between two nodes and the other for populating the tree with nodes
for test purposes.