FIX: Function for Finding Shortest distance
Changed the formula for finding the shortest distance between two nodes.
Used their common ancestor that is closest to each of them, by finding
the last index of nodes that both of their paths share.

And using the index of their common ancestor, the size of their path
is diminished by the common ancestor index.

The path is generated from the root node which is why the shortest
distance between the nodes can be located by subtracting from their
size the index of their closest common ancestor.